+++ config-templates.tf
@@ -15,8 +15,8 @@ data "template_file" "beekeeper_graphite_config" {
   }
 }
 
-data "template_file" "beekeeper_path_scheduler_config" {
-  template = file("${path.module}/files/beekeeper-path-scheduler-config.yml")
+data "template_file" "beekeeper_scheduler_apiary_config" {
+  template = file("${path.module}/files/beekeeper-scheduler-apiary-config.yml")
 
   vars = {
     db_endpoint      = aws_db_instance.beekeeper.endpoint
@@ -26,14 +26,27 @@ data "template_file" "beekeeper_path_scheduler_config" {
   }
 }
 
-data "template_file" "beekeeper_cleanup_config" {
-  template = file("${path.module}/files/beekeeper-cleanup-config.yml")
+data "template_file" "beekeeper_path_cleanup_config" {
+  template = file("${path.module}/files/beekeeper-path-cleanup-config.yml")
 
   vars = {
     db_endpoint        = aws_db_instance.beekeeper.endpoint
     db_username        = aws_db_instance.beekeeper.username
     scheduler_delay_ms = var.scheduler_delay_ms
     dry_run_enabled    = var.dry_run_enabled
+    graphite_config    = var.graphite_enabled == "false" ? "" : data.template_file.beekeeper_graphite_config.rendered
+  }
+}
+
+data "template_file" "beekeeper_metadata_cleanup_config" {
+  template = file("${path.module}/files/beekeeper-metadata-cleanup-config.yml")
+
+  vars = {
+    db_endpoint        = aws_db_instance.beekeeper.endpoint
+    db_username        = aws_db_instance.beekeeper.username
+    scheduler_delay_ms = var.scheduler_delay_ms
+    dry_run_enabled    = var.dry_run_enabled
+    metastore_uri      = var.metastore_uri
     graphite_config    = var.graphite_enabled == "false" ? "" : data.template_file.beekeeper_graphite_config.rendered
   }
 }