+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2020-09-11
+### Added
+- Kubernetes deployment options for new Metadata Cleanup module. 
+### Changed
+- Renamed `Cleanup` to `Path Cleanup`, and `Path Scheduler` to `Scheduler Apiary` to match name changes in Beekeeper.
+
 ## [2.1.3] - 2020-08-07
 ### Changed
 - Fix multi instance deployment resource names.